remove tx_msg
