+++ crates/ibc/src/core/ics03_connection/msgs/conn_open_confirm.rs
@@ -7,7 +7,7 @@ use ibc_proto::ibc::core::connection::v1::MsgConnectionOpenConfirm as RawMsgConn
 
 use crate::core::ics03_connection::error::ConnectionError;
 use crate::core::ics24_host::identifier::ConnectionId;
-use crate::core::tx_msg::Msg;
+use crate::core::Msg;
 use crate::signer::Signer;
 
 pub(crate) const TYPE_URL: &str = "/ibc.core.connection.v1.MsgConnectionOpenConfirm";
