+++ crates/ibc/src/core/ics03_connection/msgs/conn_open_init.rs
@@ -9,7 +9,7 @@ use crate::core::ics03_connection::connection::Counterparty;
 use crate::core::ics03_connection::error::ConnectionError;
 use crate::core::ics03_connection::version::Version;
 use crate::core::ics24_host::identifier::ClientId;
-use crate::core::tx_msg::Msg;
+use crate::core::Msg;
 use crate::signer::Signer;
 
 pub(crate) const TYPE_URL: &str = "/ibc.core.connection.v1.MsgConnectionOpenInit";
