+++ crates/ibc/src/core/ics04_channel/msgs/chan_close_init.rs
@@ -6,7 +6,7 @@ use ibc_proto::ibc::core::channel::v1::MsgChannelCloseInit as RawMsgChannelClose
 
 use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics24_host::identifier::{ChannelId, PortId};
-use crate::core::tx_msg::Msg;
+use crate::core::Msg;
 use crate::signer::Signer;
 
 pub(crate) const TYPE_URL: &str = "/ibc.core.channel.v1.MsgChannelCloseInit";
