+++ crates/ibc/src/core/ics04_channel/msgs/chan_open_ack.rs
@@ -2,7 +2,7 @@ use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics04_channel::Version;
 use crate::core::ics23_commitment::commitment::CommitmentProofBytes;
 use crate::core::ics24_host::identifier::{ChannelId, PortId};
-use crate::core::tx_msg::Msg;
+use crate::core::Msg;
 use crate::signer::Signer;
 use crate::{prelude::*, Height};
