+++ crates/ibc/src/core/ics04_channel/msgs/chan_open_confirm.rs
@@ -1,7 +1,7 @@
 use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics23_commitment::commitment::CommitmentProofBytes;
 use crate::core::ics24_host::identifier::{ChannelId, PortId};
-use crate::core::tx_msg::Msg;
+use crate::core::Msg;
 use crate::signer::Signer;
 use crate::{prelude::*, Height};
