+++ crates/ibc/src/core/ics04_channel/msgs/recv_packet.rs
@@ -7,7 +7,7 @@ use ibc_proto::ibc::core::channel::v1::MsgRecvPacket as RawMsgRecvPacket;
 use crate::core::ics04_channel::error::PacketError;
 use crate::core::ics04_channel::packet::Packet;
 use crate::core::ics23_commitment::commitment::CommitmentProofBytes;
-use crate::core::tx_msg::Msg;
+use crate::core::Msg;
 use crate::signer::Signer;
 use crate::Height;
